Moved search path functionality to its package
